Combined name and genc transformers.
